app/requirements.txt
asyncio
cogwatch
debugpy
discord.py
PyNaCl
yt-dlp
